requirements.txt
tensorflow==2.4.1
tensorflow-datasets==4.0.1
gin-config==0.4.0
tensorflow-addons==0.11.2
matplotlib==3.3.2
seaborn==0.11.0
prettytable==2.0.0

